backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
pydantic-settings
passlib[bcrypt]
python-jose
PyJWT
telethon
fastapi-csrf-protect